+++ public/index.php
@@ -3,6 +3,7 @@
  * The Turbine Micro Framework. An advanced derivate of Proton Micro Framework
  *
  * @author Marco Bunge <marco_bunge@web.de>
+ * @author Alex Bilbie <hello@alexbilbie.com>
  * @copyright Marco Bunge <marco_bunge@web.de>
  *
  * @license MIT
@@ -10,7 +11,7 @@
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use Turbine\Application;
+use ZeroXF10\Turbine\Application;
 
 require_once __DIR__ . '/../vendor/autoload.php';