+++ src/Application/ApplicationEvent.php
@@ -9,13 +9,13 @@
  * @license MIT
  */
 
-namespace Turbine\Application;
+namespace ZeroXF10\Turbine\Application;
 
 
 use League\Event\AbstractEvent;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use Turbine\ApplicationInterface;
+use ZeroXF10\Turbine\ApplicationInterface;
 use Zend\Stdlib\ArrayObject;
 
 class ApplicationEvent extends AbstractEvent