+++ src/Symfony/HttpKernelAdapter.php
@@ -9,11 +9,11 @@
  * @license MIT
  */
 
-namespace Turbine\Symfony;
+namespace ZeroXF10\Turbine\Symfony;
 
 
 
-use Turbine\Application;
+use ZeroXF10\Turbine\Application;
 use Symfony\Bridge\PsrHttpMessage\Factory\DiactorosFactory;
 use Symfony\Bridge\PsrHttpMessage\Factory\HttpFoundationFactory;
 use Symfony\Component\HttpFoundation\Request;