+++ tests/Symfony/HttpKernelAdapterTest.php
@@ -9,12 +9,12 @@
  * @license MIT
  */
 
-namespace Turbine\Tests\Symfony;
+namespace ZeroXF10\Turbine\Tests\Symfony;
 
 
 use League\Route\Http\Exception\NotFoundException;
-use Turbine\Application;
-use Turbine\Symfony\HttpKernelAdapter;
+use ZeroXF10\Turbine\Application;
+use ZeroXF10\Turbine\Symfony\HttpKernelAdapter;
 use Psr\Http\Message\ResponseInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\HttpKernelInterface;